-fixed game boots if you just press a and the pointer is off screen (issue 125) -set the button type to s16 (now its correct :P) -corrected some possible bugs with the button gui management

+++ trunk/source/gui/gui.hpp
@@ -30,26 +30,26 @@ public:
 	bool init(CVideo &vid);
 	void setRumble(bool enabled) { m_rumbleEnabled = enabled; }
 	void reserve(u32 capacity) { m_elts.reserve(capacity); }
-	u16 addButton(SFont font, const wstringEx &text, int x, int y, u32 width, u32 height, const CColor &color,
+	s16 addButton(SFont font, const wstringEx &text, int x, int y, u32 width, u32 height, const CColor &color,
 		const SButtonTextureSet &texSet, const SmartGuiSound &clickSound = _noSound, const SmartGuiSound &hoverSound = _noSound);
-	u16 addLabel(SFont font, const wstringEx &text, int x, int y, u32 width, u32 height, const CColor &color, u16 style, const STexture &bg = _noTexture);
-	u16 addPicButton(const u8 *pngNormal, const u8 *pngSelected, int x, int y, u32 width, u32 height,
+	s16 addLabel(SFont font, const wstringEx &text, int x, int y, u32 width, u32 height, const CColor &color, s16 style, const STexture &bg = _noTexture);
+	s16 addPicButton(const u8 *pngNormal, const u8 *pngSelected, int x, int y, u32 width, u32 height,
 		const SmartGuiSound &clickSound = _noSound, const SmartGuiSound &hoverSound = _noSound);
-	u16 addPicButton(STexture &texNormal, STexture &texSelected, int x, int y, u32 width, u32 height,
+	s16 addPicButton(STexture &texNormal, STexture &texSelected, int x, int y, u32 width, u32 height,
 		const SmartGuiSound &clickSound = _noSound, const SmartGuiSound &hoverSound = _noSound);
-	u16 addProgressBar(int x, int y, u32 width, u32 height, SButtonTextureSet &texSet);
-	void setText(u16 id, const wstringEx &text, bool unwrap = false);
-	void setText(u16 id, const wstringEx &text, u32 startline, bool unwrap = false);
-	void setBtnTexture(u16 id, STexture &texNormal, STexture &texSelected);
-	void setTexture(u16 id ,STexture &bg);
-	void setTexture(u16 id, STexture &bg, int width, int height);
-	void setProgress(u16 id, float f, bool instant = false);
-	void reset(u16 id, bool instant = false);
-	void moveBy(u16 id, int x, int y, bool instant = false);
-	void getDimensions(u16 id, int &x, int &y, u32 &width, u32 &height);
-	void hide(u16 id, int dx, int dy, float scaleX, float scaleY, bool instant = false);
-	void hide(u16 id, bool instant = false);
-	void show(u16 id, bool instant = false);
+	s16 addProgressBar(int x, int y, u32 width, u32 height, SButtonTextureSet &texSet);
+	void setText(s16 id, const wstringEx &text, bool unwrap = false);
+	void setText(s16 id, const wstringEx &text, u32 startline, bool unwrap = false);
+	void setBtnTexture(s16 id, STexture &texNormal, STexture &texSelected);
+	void setTexture(s16 id ,STexture &bg);
+	void setTexture(s16 id, STexture &bg, int width, int height);
+	void setProgress(s16 id, float f, bool instant = false);
+	void reset(s16 id, bool instant = false);
+	void moveBy(s16 id, int x, int y, bool instant = false);
+	void getDimensions(s16 id, int &x, int &y, u32 &width, u32 &height);
+	void hide(s16 id, int dx, int dy, float scaleX, float scaleY, bool instant = false);
+	void hide(s16 id, bool instant = false);
+	void show(s16 id, bool instant = false);
 	void mouse(int chan, int x, int y);
 	void up(void);
 	void down(void);
@@ -57,10 +57,10 @@ public:
 	void tick(void);
 	void noClick(bool noclick = false);
 	void noHover(bool nohover = false);
-	void click(u16 id = (u32)-1);
-	bool selected(u16 button = (u32)-1);
+	void click(s16 id = -1);
+	bool selected(s16 button = -1);
 	void setRumble(int, bool wii = false, bool gc = false);
-	void deselect(void){ for(int chan = WPAD_MAX_WIIMOTES-1; chan >= 0; chan--) m_selected[chan] = (u32)-1; }
+	void deselect(void){ for(int chan = WPAD_MAX_WIIMOTES-1; chan >= 0; chan--) m_selected[chan] = -1; }
 	void stopSounds(void);
 	void setSoundVolume(int vol);
 private:
@@ -138,7 +138,7 @@ private:
 	};
 private:
 	vector<SmartPtr<SElement> > m_elts;
-	u32 m_selected[WPAD_MAX_WIIMOTES];
+	s32 m_selected[WPAD_MAX_WIIMOTES];
 	bool m_rumbleEnabled;
 	u8 m_rumble[WPAD_MAX_WIIMOTES];
 	bool wii_rumble[WPAD_MAX_WIIMOTES];
