+++ trunk/source/menu/menu_config4.cpp
@@ -43,7 +43,7 @@ void CMenu::_hideConfig4(bool instant)
 	m_btnMgr.hide(m_config4BtnReturnToM, instant);
 	m_btnMgr.hide(m_config4BtnReturnToP, instant);
 	for(u8 i = 0; i < ARRAY_SIZE(m_config4LblUser); ++i)
-		if(m_config4LblUser[i] != (u16)-1)
+		if(m_config4LblUser[i] != -1)
 			m_btnMgr.hide(m_config4LblUser[i], instant);
 }
 
@@ -63,7 +63,7 @@ void CMenu::_showConfig4(void)
 	m_btnMgr.show(m_config4BtnReturnToP);
 
 	for(u32 i = 0; i < ARRAY_SIZE(m_config4LblUser); ++i)
-		if(m_config4LblUser[i] != (u16)-1)
+		if(m_config4LblUser[i] != -1)
 			m_btnMgr.show(m_config4LblUser[i]);
  
 	int i;
